pyproject.toml
[build-system]
requires = ["setuptools==80.9.0", "wheel==0.45.1"]
build-backend = "setuptools.build_meta"

[project]
name = "controldiff"
version = "0.1.0"
description = "Agentic regulatory change management for fintech compliance teams."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "ControlDiff" }
]
dependencies = [
  "alembic==1.18.4",
  "fastapi==0.135.1",
  "httpx==0.28.1",
  "jinja2==3.1.6",
  "langfuse==4.0.0",
  "langgraph==1.0.10",
  "openai==2.26.0",
  "orjson==3.11.7",
  "prometheus-client==0.24.1",
  "psycopg[binary]==3.3.3",
  "pydantic==2.12.5",
  "pydantic-settings==2.13.1",
  "python-dotenv==1.1.1",
  "python-multipart==0.0.20",
  "qdrant-client==1.17.0",
  "sqlalchemy==2.0.48",
  "uvicorn==0.41.0",
]

[project.optional-dependencies]
dev = [
  "pytest==8.4.0",
  "pytest-asyncio==1.3.0",
  "pytest-cov==6.0.0",
  "ruff==0.15.5",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src", "tests", "scripts"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
